refactor: use concurrent go-spew config
Instead of modifying the global state and restoring it, use the
concurrent configuration object to configure go-spew. This removes the
need for defer, and makes tests with snapshot marginally safer if tests
are run concurrently.

Mostly though, it's a little cleaner, and makes it more straightforward to
add more configuration.