fix: use stable key ordering for snapshots with maps
Snapshots containing maps are unstable because iteration order of a map's
keys is not guaranteed.

The spew library can sort map keys that are sortable to provide
stability at the cost of speed. Some types are not sortable, and there
is a spew option to deal with this.

This option first serializes keys to strings using spew, then sorts.
This behaviour should probably be opt-in, so it's not included here.